requirements.txt
graphviz
pandas
snowflake-snowpark-python
streamlit
streamlit-aggrid